+++ zcash_client_backend/src/data_api/error.rs
@@ -12,6 +12,7 @@ use zcash_protocol::{
     value::{BalanceError, Zatoshis},
 };
 
+use crate::wallet::OutputRef;
 use crate::{
     data_api::wallet::input_selection::InputSelectorError, fees::ChangeError,
     proposal::ProposalError, wallet::NoteId,
@@ -441,5 +442,37 @@ impl<DE, TE, SE, FE, CE, N> From<pczt::roles::tx_extractor::Error>
 {
     fn from(e: pczt::roles::tx_extractor::Error) -> Self {
         Error::Pczt(PcztError::Extraction(e))
+    }
+}
+
+/// Errors that occur when attempting to lock an output.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum LockError<S> {
+    /// Wrapper for storage errors.
+    Storage(S),
+    /// The wrapped output reference was not found, or the output it refers to was already locked.
+    LockFailure(OutputRef),
+}
+
+impl<S: Display> Display for LockError<S> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            LockError::Storage(e) => write!(f, "Note locking failed: {e}"),
+            LockError::LockFailure(output) => {
+                write!(
+                    f,
+                    "Lock conflict or missing output for reference {output:?}"
+                )
+            }
+        }
+    }
+}
+
+impl<S: std::error::Error> std::error::Error for LockError<S> {
+    fn cause(&self) -> Option<&dyn std::error::Error> {
+        match self {
+            LockError::Storage(e) => Some(e),
+            LockError::LockFailure(_) => None,
+        }
     }
 }