+++ zcash_client_backend/CHANGELOG.md
@@ -32,6 +32,7 @@ workspace.
   - `impl From<sapling_crypto::Note> for Note`
   - `impl From<orchard::Note> for Note`
 - `zcash_client_backend::data_api::ll`
+- `zcash_client_backend::data_api::error::LockError`
 - `zcash_client_backend::wallet::OutputRef`
 - `impl From<zcash_client_backend::wallet::NoteId> for zcash_client_backend::wallet::OutputRef`
 
@@ -77,7 +78,10 @@ workspace.
   and `IncompatibleTxVersion`
 - `zcash_client_backend::proposal::ProposalError::ChainDoubleSpend` now wraps an
   `OutputRef` instead of `(PoolType, TxId, u32)`.
-- `zcash_client_backend::data_api::WalletWrite` has added method `truncate_to_chain_state`.
+- `zcash_client_backend::data_api::WalletWrite` has added methods
+  `truncate_to_chain_state`, `lock_outputs` and `unlock_output`.
+- `zcash_client_backend::data_api::WalletTest` has added method
+  `get_locked_outputs`.
 - The associated type `zcash_client_backend::fees::ChangeStrategy::MetaSource` is now
   bounded on the newly added `MetaSource` type instead of
   `zcash_client_backend::data_api::InputSource`.